[pci] Shift to using mx_pci_resource_t for memory
PCI syscalls and protocols have been added to get pci_resource_t objects
that encapsulate MMIO and PIO resources (Configs and BARs). In the MMIO
case this means we're passing VMOs from the bus driver to users of the
protocol. PIO is platform defined, but on x86 corresponds to an addr /
size range.

Change-Id: Ibcbd695efe3900316949adb837ee3c6a690c5802