[mx] Add MX_CACHE_POLICY_USE_EXISTING
Add an additional enum value for cache to handle cases where an api
may need to take a cache policy, but the caller doesn't want to change
how the object is currently configured. Since MX_CACHE_POLICY_PAGED
defaulted to zero it was not suitable as a default argument in these
cases. Adding MX_CACHE_POLICY_USE_EXISTING also ensures that the bits
still line with ARCH_MMU_FLAGS and is caught as invalid by the
sys_mmap_device_memory syscall.

Change-Id: I97cc52d2a0f15f723e1837d98efc044d90eeb443